pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[project]
authors = [
    {name = "Detlef Wuppermann", email = "detlef@coding-on-my-mind.com"},
]

name = "context-configuration"
dynamic = ["version"]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
description = "Configuration management based on profiles and annotations."
requires-python = ">=3.9"

dependencies = [
    "PyYAML >= 6.0.0",
]

keywords = ["configuration", "context", "yaml", "cli", "envvars", "bean"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
]

[project.urls]
homepage = "https://coding-on-my-mind.com"
repository = "https://github.com/coding-on-my-mind/context-configuration"
issues = "https://github.com/coding-on-my-mind/context-configuration/issues"

[project.optional-dependencies]
dev = [
    "pytest == 7.4",
    "pytest-mock == 3.11.1",
    "pytest-cov == 4.1.0",
    "coverage == 7.4.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
version = {file = "__version__.txt"}

[tool.pytest.ini_options]
pythonpath = [
    "./src"
]
addopts = [
    "--import-mode=importlib",
]
